backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
rdkit-pypi==2022.9.5
aizynthfinder==4.2.1
pulp==2.9.0
networkx==3.3
numpy==1.26.4
scipy==1.13.1
pydantic==2.8.2
python-multipart==0.0.9
transformers>=4.40.0
torch>=2.0.0
pytest==8.3.2
pytest-asyncio==0.23.8
